+++ src/transient/Quiet.h
@@ -80,7 +80,7 @@ class Quiet
 
     bool detect(
       const vector<float>& samples,
-      const double sampleRate,
+      const float sampleRate,
       const bool fromBackFlag,
       Interval& available);
